app/requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
idna==3.10
msgpack==1.1.0
pymemcache==3.5.2
pytz==2025.2
redis==6.2.0
requests==2.32.3
six==1.17.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
psycopg==3.2.9
psycopg-binary==3.2.9